requirements.txt
torch>=2.0.0
transformers>=4.36.0
prettytable>=2.0.0
numpy>=1.20.0
pillow>=9.0.0
accelerate>=0.20.0 